[move] Connection Dialogのディレクトリ変更
