+++ src/pages/Connection/ConnectionDetail/ConnectionDialog.module.scss
